requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14,<4.0
Faker>=18.0,<19.0
